// aries-indeed_framework/backend/src/generic.modules/database/db.models.enum.ts
export enum databaseModels {
    LoggedDataDBModel = 'LoggedDataModel',
    RheinenergieDataDBModel = 'RheinenergieDataModel',
    SMADataDBModel = 'SMADataModel',
    SMAprobandenDataDBModel = 'SMAprobandenDataDB',
    LiqwotecDataDBModel = 'LiqwotecDataDBModel',
    LEWDataDBModel = 'LEWDataDBModel',
    SchweigerDataDBModel = 'SchweigerDataDBModel',
    ffeDataDBModel = 'ffeDataDBModel',
    AggregatedDataDBModel = 'AggregatedDataModel',
    NotarizedDataDBModel = 'NotarizedDataModel',


    NotarizedOwnerDataDBModel = 'NotarizedOwnerDataDBModel',
    RemoteNotarizationDataDBModel = 'RemoteNotarizationDataDBModel',
    CollectiveNotarizationDataDBModel = 'CollectiveNotarizationDataDBModel',

    LoggedProsumerDataDBModel = 'LoggedProsumerDataModel',
    CommittedProsumerDataDBModel = 'CommittedProsumerDataModel',
    VerifiableProsumerDataDBModel = 'VerifiableProsumerDataModel',
    labeledConsumerAggregationDBModel = 'labeledConsumerAggregationDBModel',

    ConsumerConsumptionDBModel = 'ConsumerConsumptionModel',
    ConsumerAggregationDBModel = 'ConsumerAggregationModel',
    ConsumerCompilationDBModel = 'ConsumerCompilationModel',
    ConsumerCompilationAggregationDBModel = 'ConsumerCompilationAggregationModel',
    ConsumerFootprintDBModel = 'ConsumerFootprintModel',
    ConsumerFootprintDataDBModel = 'ConsumerFootprintDataModel',
    ConsumerForecastDBModel = 'ConsumerForecastModel',
    ConsumerPriorizationDBModel = 'ConsumerPriorizationModel',
    masterRegistryDBModel = 'masterRegistryModel',
    usersDBModel = 'usersDBModel'
}