// aries-indeed_framework/backend/src/generic.modules/schemas/schemas.enum.ts
export enum dbschemas {
    LoggedData = 'LoggedData',
    AggregatedData = 'AggregatedData',
    NotarizedData = 'NotarizedData',
    NotarizedOwnerData = 'NotarizedOwnerDate',
    RemoteNotarizationData ='RemoteNotarizationData',
    CollectiveNotarizationData ='CollectiveNotarizationData',
    RheinenergieData = 'RheinenergieData',
    SMAData = 'SMAData',
    LiqwotecData = 'LiqwotecData',
    LEWData = 'LEWData',
    SchweigerData = 'SchweigerData',
    LoggedProsumerData = 'LoggedProsumerData',
    CommittedProsumerData = 'CommittedProsumerData',
    VerifiableProsumerData = 'VerifiableProsumerData',
    labeledConsumerAggregation = 'labeledConsumerAggregation',

    ConsumerConsumption = 'ConsumerConsumption',
    ConsumerAggregation = 'ConsumerAggregation',
    ConsumerCompilation = 'ConsumerCompilation',
    ConsumerCompilationAggregation = 'ConsumerCompilationAggregation',
    ConsumerFootprint = 'ConsumerFootprint',
    ConsumerFootprintData = 'ConsumerFootprintData',
    ConsumerForecast = 'ConsumerForecast',
    ConsumerPriorization = 'ConsumerPriorization',
    masterRegistry = 'masterRegistry',
    ffeData = 'ffeData',
    user = 'User'
}